[CDS-44756]: Remove dependency between NG Manager and Pipeline
