+++ 125-cd-nextgen/src/main/java/io/harness/cdng/ssh/output/SshInfraDelegateConfigOutput.java
@@ -5,22 +5,22 @@
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-package io.harness.steps.shellscript;
+package io.harness.cdng.ssh.output;
 
 import static io.harness.annotations.dev.HarnessTeam.CDP;
 
 import io.harness.annotation.RecasterAlias;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.delegate.task.ssh.SshInfraDelegateConfig;
 import io.harness.pms.sdk.core.data.ExecutionSweepingOutput;
 
-import java.util.Set;
 import lombok.Builder;
 import lombok.Value;
 
 @Value
 @Builder
 @OwnedBy(CDP)
-@RecasterAlias("io.harness.steps.shellscript.HostsOutput")
-public class HostsOutput implements ExecutionSweepingOutput {
-  Set<String> hosts;
+@RecasterAlias("io.harness.cdng.ssh.output.SshInfraDelegateConfigOutput")
+public class SshInfraDelegateConfigOutput implements ExecutionSweepingOutput {
+  SshInfraDelegateConfig sshInfraDelegateConfig;
 }
