+++ 125-cd-nextgen/src/main/java/io/harness/cdng/ssh/output/WinRmInfraDelegateConfigOutput.java
@@ -5,13 +5,13 @@
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-package io.harness.steps.shellscript;
+package io.harness.cdng.ssh.output;
 
 import static io.harness.annotations.dev.HarnessTeam.CDP;
 
 import io.harness.annotation.RecasterAlias;
 import io.harness.annotations.dev.OwnedBy;
-import io.harness.delegate.task.ssh.SshInfraDelegateConfig;
+import io.harness.delegate.task.ssh.WinRmInfraDelegateConfig;
 import io.harness.pms.sdk.core.data.ExecutionSweepingOutput;
 
 import lombok.Builder;
@@ -20,7 +20,7 @@ import lombok.Value;
 @Value
 @Builder
 @OwnedBy(CDP)
-@RecasterAlias("io.harness.steps.shellscript.SshInfraDelegateConfigOutput")
-public class SshInfraDelegateConfigOutput implements ExecutionSweepingOutput {
-  SshInfraDelegateConfig sshInfraDelegateConfig;
+@RecasterAlias("io.harness.cdng.ssh.output.WinRmInfraDelegateConfigOutput")
+public class WinRmInfraDelegateConfigOutput implements ExecutionSweepingOutput {
+  WinRmInfraDelegateConfig winRmInfraDelegateConfig;
 }
