+++ 125-cd-nextgen/src/test/resources/recaster/alias/alias.txt
@@ -122,9 +122,9 @@ io.harness.cdng.artifact.bean.yaml.AMIArtifactConfig
 io.harness.cdng.artifact.bean.yaml.ami.AMITag
 io.harness.cdng.artifact.bean.yaml.ami.AMIFilter
 io.harness.steps.shellscript.K8sInfraDelegateConfigOutput
-io.harness.steps.shellscript.SshInfraDelegateConfigOutput
-io.harness.steps.shellscript.WinRmInfraDelegateConfigOutput
-io.harness.steps.shellscript.HostsOutput
+io.harness.cdng.ssh.output.SshInfraDelegateConfigOutput
+io.harness.cdng.ssh.output.WinRmInfraDelegateConfigOutput
+io.harness.cdng.ssh.output.HostsOutput
 io.harness.cdng.infra.beans.K8sGcpInfraMapping
 io.harness.cdng.infra.beans.K8sAzureInfraMapping
 io.harness.cdng.infra.beans.PdcInfraMapping
@@ -527,3 +527,6 @@ io.harness.cdng.gitops.UpdateReleaseRepoOutcome
 io.harness.cdng.instance.outcome.InstancesOutcome
 io.harness.cdng.instance.outcome.InstanceOutcome
 io.harness.cdng.instance.outcome.HostOutcome
+io.harness.cdng.ssh.output.HostsOutput
+io.harness.cdng.ssh.output.SshInfraDelegateConfigOutput
+io.harness.cdng.ssh.output.WinRmInfraDelegateConfigOutput
