+++ 300-cv-nextgen/src/test/resources/recaster/alias/alias.txt
@@ -37,8 +37,6 @@ io.harness.pms.sdk.core.plan.MapStepParameters
 io.harness.steps.jira.create.JiraCreateSpecParameters
 io.harness.steps.common.pipeline.PipelineSetupStepParameters
 io.harness.pms.yaml.validation.InputSetValidator
-io.harness.steps.shellscript.SshInfraDelegateConfigOutput
-io.harness.steps.shellscript.WinRmInfraDelegateConfigOutput
 io.harness.steps.barriers.beans.BarrierOutcome
 io.harness.when.beans.StageWhenCondition
 io.harness.steps.approval.step.jira.JiraApprovalSpecParameters
