+++ pipeline-service/modules/orchestration-steps/src/test/resources/recaster/alias/alias.txt
@@ -40,9 +40,6 @@ io.harness.steps.common.pipeline.PipelineSetupStepParameters
 io.harness.pms.yaml.validation.InputSetValidator
 io.harness.pms.pipeline.stage.PipelineStageNode
 io.harness.steps.shellscript.K8sInfraDelegateConfigOutput
-io.harness.steps.shellscript.SshInfraDelegateConfigOutput
-io.harness.steps.shellscript.WinRmInfraDelegateConfigOutput
-io.harness.steps.shellscript.HostsOutput
 io.harness.steps.barriers.beans.BarrierOutcome
 io.harness.when.beans.StageWhenCondition
 io.harness.steps.approval.step.jira.JiraApprovalSpecParameters
