+++ pipeline-service/service/src/test/resources/recaster/alias/alias.txt
@@ -8,9 +8,6 @@ io.harness.plancreator.steps.internal.FlagConfigurationStepInfo
 io.harness.steps.jira.update.beans.TransitionTo
 barrierSpecParameters
 io.harness.steps.shellscript.K8sInfraDelegateConfigOutput
-io.harness.steps.shellscript.SshInfraDelegateConfigOutput
-io.harness.steps.shellscript.WinRmInfraDelegateConfigOutput
-io.harness.steps.shellscript.HostsOutput
 io.harness.steps.common.pipeline.PipelineSetupStepParameters
 io.harness.when.beans.StepWhenCondition
 io.harness.plancreator.steps.common.StepElementParameters
